+++ internal/provider/resource_datapower_wsrrsavedsearchsubscription.go
@@ -139,7 +139,7 @@ func (r *WSRRSavedSearchSubscriptionResource) Create(ctx context.Context, req re
 	body := data.ToBody(ctx, `WSRRSavedSearchSubscription`)
 	res, err := r.client.Post(data.GetPath(), body)
 
-	if err != nil && res.RawResponse.StatusCode != 409 {
+	if err != nil && !strings.Contains(err.Error(), "status 409") {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to create object (%s), got error: %s, %s", "POST", err, res.String()))
 		return
 	}
@@ -160,7 +160,7 @@ func (r *WSRRSavedSearchSubscriptionResource) Read(ctx context.Context, req reso
 		return
 	}
 	res, err := r.client.Get(data.GetPath() + "/" + data.Id.ValueString())
-	if err != nil && (strings.Contains(err.Error(), "StatusCode 404") || strings.Contains(err.Error(), "StatusCode 406") || strings.Contains(err.Error(), "StatusCode 500") || strings.Contains(err.Error(), "StatusCode 400")) {
+	if err != nil && (strings.Contains(err.Error(), "status 404") || strings.Contains(err.Error(), "status 406") || strings.Contains(err.Error(), "status 500") || strings.Contains(err.Error(), "status 400")) {
 		resp.State.RemoveResource(ctx)
 		return
 	} else if err != nil {
@@ -208,7 +208,7 @@ func (r *WSRRSavedSearchSubscriptionResource) Delete(ctx context.Context, req re
 		return
 	}
 	res, err := r.client.Delete(data.GetPath() + "/" + data.Id.ValueString())
-	if err != nil && res.RawResponse.StatusCode != 404 && res.RawResponse.StatusCode != 409 {
+	if err != nil && !strings.Contains(err.Error(), "status 404") && !strings.Contains(err.Error(), "status 409") {
 		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object (DELETE), got error: %s, %s", err, res.String()))
 		return
 	}